libcommon: IWorkItem and WorkQueue
IWorkItem is an interface to inherit if you wish to use
your object with a WorkQueue. Simply queue up your void(void)
function objects with the enqueue function and then call
the WorkQueue's enqueue function with your object.

WorkQueue is an object which hosts multiple threads (as
specified by the constructor) to do work.