libcommon: Track if WorkItem is on WorkQueue
This commit modifies WorkItem to track whether the item is queued
up on a WorkQueue; the WorkItem can only be enqueued onto its
WorkQueue once. When the job spools, if the queue is nonempty,
then the item is once again queued up onto its WorkQueue. Otherwise,
the enqueue function will enqueue the WorkItem when its list of
deferred functions goes from being empty to nonempty.