libnl: NetlinkServer send() function
This commit introduces a send function to NetlinkServer. This
function creates a socket on which to send the message (note
that if you are already listening to the same type and family,
this call will fail and return false) and sends the message,
then awaits the response. It will invoke the supplied callback
once for each received message on the socket and clean up
once the sentinel message (NLMSG_DONE) is received.

Beware that as a consequence of this change, callbacks will
be invoked for NLMSG_DONE message types.