libnl: NetlinkServer's send() function to call callback with DONE
When the NetlinkServer receives the NLMSG_DONE message, it previously
would ignore the family and group associated with the message and not
call the callback function. This can leave the caller hanging since
it has no way of knowing when the stream of messages will end. This
change causes the NetlinkServer to call back with the DONE message.
Note also that the callback needs to happen BEFORE the call to
ignore, since ignore will release the callback function, causing
some strange behaviour.