libnet: NetworkInterfaceStats class
This commit introduces a NetworkInterfaceStats class. This class
is mainly meant to make a copy of the underlying rtnl_link_stats
which is assumed to belong to some buffer::Buffer. The idea is
that we may want to keep the statistics around (such as in a
statistics collection class) without keeping the entire heap-
allocated buffer around.

This commit also adds the interface index to the NetworkInterface
class.