libnet: NetworkInterfaceStatsMonitor
This commit introduces the NetworkInterfaceStatsMonitor class
and some related classes. This class can be configured to
monitor the network interface statistics for a particular
interface, with a specified sample frequency and sample
size. The caller will be called back every time the specified
number of samples has been collected.

This works, but it could be a little more polished. For example,
the NetworkInterfaceStatsBuffer should not directly expose the
rtnl_link_stats structure. Moreover, the monitor class should
accept a secondary WorkQueue into which it can defer waiting
jobs, since they will block the (single-threaded) NetlinkServer,
thus making it useless for other callers.