NetlinkServer: transmit function
This commit factors out the main part of the 'send' function,
which also expects a callback, to create a 'transmit' function.
Given a UUID, this function will find the corresponding socket
and transmit the argument message. This pattern is meant to be
used for sockets on which you will send and receive multiple
message sequences, as opposed to the send function, which
expects to see a sequence of messages terminated by a NLMSG_DONE
type message.