libnl: 'Expect done' argument to NetlinkServer::send
At present, the NetlinkServer::send function expects to receive
a sequence of netlink messages, followed by a terminating NLMSG_
DONE message. If it does not receive the final NLMSG_DONE, the
corresponding netlink socket is not properly cleaned up, and other
bad behaviours can ensue. This commit addresses the issue by
allowing the caller to specify whether the function should expect
a NLMSG_DONE message. If so, the behaviour is unchanged. If not,
then the socket is closed after the receipt of the first message.