libnl: NetlinkMessageHeader class
Instead of using struct nlmsghdr, this commit introduces a class
which represents the header. It is still possible to get access
to the raw pointer if necessary, but it is preferred to use this
class. This causes a few cascading changes in other classes, which
are also included in the commit.