libnl: NetlinkSocket 'allocate' function
Prior to this commit, the NetlinkSocket object constructor accepted
a single argument, namely the netlink family. It would then attempt
to create a socket in its constructor. This meant the constructor
could fail and have no way to communicate that failure back to the
caller. This commit updates the NetlinkSocket class to create the
socket in an 'allocate' function which can return false to indicate
a failure.